+++ src/components/booklet-image/__tests__/booklet-image-test-cases.js
@@ -1,0 +1,13 @@
+import BookletImage from '../booklet-image';
+
+const testCases = {};
+
+testCases.basic = {
+  component: BookletImage,
+  description: 'Basic',
+  props: {
+    size: '60'
+  }
+};
+
+export { testCases };