+++ src/components/product-menu/__tests__/product-menu-test-cases.js
@@ -53,4 +53,4 @@ testCases.beta = {
   }
 };
 
-export default { testCases };
+export { testCases };