+++ src/components/search/search-box.js
@@ -41,6 +41,7 @@ class SearchBox extends React.Component {
   }
 
   closeModal() {
+    this.props.reset();
     this.setState({ modalOpen: false });
   }
 
@@ -59,32 +60,38 @@ class SearchBox extends React.Component {
       <div className="py12 border-b border--gray-faint mx6">
         <div className="toggle-group">
           <div className="toggle-container">
-            <a
+            <button
               key={getFilterValueDisplay(siteFilter.value)}
               className={`toggle py3 ${
                 value === siteFilter.value ? 'bg-gray color-white' : ''
               }`}
-              href="/"
               onClick={e => {
                 e.preventDefault();
+                // track click
+                if (window && window.analytics) {
+                  analytics.track('Searched docs', {
+                    query: this.props.searchTerm,
+                    toggle: true,
+                    site: this.props.site
+                  });
+                }
                 onSelect(siteFilter.value);
               }}
             >
               {getFilterValueDisplay(siteFilter.value)}
-            </a>
+            </button>
           </div>
 
           <div className="toggle-container">
-            <a
+            <button
               onClick={e => {
                 e.preventDefault();
                 onRemove(value);
               }}
               className={`toggle py3 ${!value ? 'bg-gray color-white' : ''}`}
-              href="/"
             >
               All docs
-            </a>
+            </button>
           </div>
         </div>
       </div>
@@ -100,12 +107,25 @@ class SearchBox extends React.Component {
         id={this.props.inputId}
         inputValue={this.props.searchTerm}
         onChange={selection => {
+          // track click
+          if (window && window.analytics) {
+            analytics.track('Searched docs', {
+              query: this.props.searchTerm,
+              clicked: selection.url.raw
+            });
+          }
           this.props.trackClickThrough(selection.id.raw); // track selection click through
           window.open(selection.url.raw, '_self'); // open selection in current window
         }}
         onInputValueChange={newValue => {
           if (props.searchTerm === newValue) return;
-          props.setSearchTerm(newValue, { debounce: 300 });
+          props.setSearchTerm(newValue, { debounce: 1500 });
+          // track query
+          if (window && window.analytics) {
+            analytics.track('Searched docs', {
+              query: newValue
+            });
+          }
         }}
         itemToString={() => props.searchTerm}
       >
@@ -134,8 +154,17 @@ class SearchBox extends React.Component {
                     <use xlinkHref="#icon-search" />
                   </svg>
                 </div>
+                {this.props.isLoading && (
+                  <div
+                    className="loading loading--s absolute bg-white"
+                    style={{
+                      top: this.state.useModal ? '21px' : '10px',
+                      right: '26px',
+                      zIndex: 5
+                    }}
+                  />
+                )}
               </label>
-
               <input
                 ref={input => {
                   this.docsSeachInput = input;
@@ -150,39 +179,40 @@ class SearchBox extends React.Component {
                   }
                 })}
               />
-              {isOpen && props.searchTerm && props.wasSearched && (
+              {isOpen && props.searchTerm && (
                 <div className="color-text shadow-darken25 round mt3 bg-white scroll-auto scroll-styled hmax360 absolute z4 w-full align-l">
                   <div>
-                    {this.props.site && (
-                      <Facet
-                        show={20}
-                        field="site"
-                        label="Site"
-                        view={this.singleLinksFacet}
-                      />
-                    )}
-                    {this.props.results.length ? (
-                      <ul>
-                        {this.props.results.map((result, index) => (
-                          <SearchResult
-                            key={index}
-                            result={result}
-                            index={index}
-                            downshiftProps={downshiftProps}
-                          />
-                        ))}
-                      </ul>
-                    ) : (
-                      <div className="py12 px12 prose">
-                        <p>
-                          Hmmm, we didn't find anything. Reword your search, or{' '}
-                          <a href="https://support.mapbox.com/hc/en-us">
-                            contact Support
-                          </a>
-                          .
-                        </p>
-                      </div>
-                    )}
+                    <Facet
+                      show={20}
+                      field="site"
+                      label="Site"
+                      view={this.singleLinksFacet}
+                    />
+
+                    {props.wasSearched &&
+                      (this.props.results.length ? (
+                        <ul>
+                          {this.props.results.map((result, index) => (
+                            <SearchResult
+                              key={index}
+                              result={result}
+                              index={index}
+                              downshiftProps={downshiftProps}
+                            />
+                          ))}
+                        </ul>
+                      ) : (
+                        <div className="py12 px12 prose">
+                          <p>
+                            Hmmm, we didn't find anything. Reword your search,
+                            or{' '}
+                            <a href="https://support.mapbox.com/hc/en-us">
+                              contact Support
+                            </a>
+                            .
+                          </p>
+                        </div>
+                      ))}
                   </div>
                 </div>
               )}
@@ -243,7 +273,9 @@ class SearchBox extends React.Component {
 SearchBox.propTypes = {
   searchTerm: PropTypes.string,
   trackClickThrough: PropTypes.func,
+  isLoading: PropTypes.bool,
   setSearchTerm: PropTypes.func,
+  reset: PropTypes.func,
   results: PropTypes.array,
   placeholder: PropTypes.string,
   background: PropTypes.oneOf(['light', 'dark']).isRequired,