+++ src/components/search/search-result.js
@@ -14,7 +14,6 @@ class SearchResult extends React.Component {
 
   render() {
     const { props } = this;
-
     const getItemProps = props.downshiftProps.getItemProps;
     const highlighted = props.downshiftProps.highlightedIndex === props.index;
     const site = this.returnRaw(props.result.site);