+++ src/components/troubleshoot-image/__tests__/troubleshoot-image-test-cases.js
@@ -1,0 +1,13 @@
+import TroubleshootImage from '../troubleshoot-image';
+
+const testCases = {};
+
+testCases.basic = {
+  component: TroubleshootImage,
+  description: 'Basic',
+  props: {
+    size: '60'
+  }
+};
+
+export { testCases };