+++ src/components/warning-image/__tests__/warning-image-test-cases.js
@@ -1,0 +1,13 @@
+import WarningImage from '../warning-image';
+
+const testCases = {};
+
+testCases.basic = {
+  component: WarningImage,
+  description: 'Basic',
+  props: {
+    size: '60'
+  }
+};
+
+export { testCases };