+++ CHANGELOG.md
@@ -1,7 +1,28 @@
-## Master
+## 0.16.2
 
-* Fixes z-index issue with `Search` modal and `TopbarSticker`. [#139](https://github.com/mapbox/dr-ui/pull/139)
+* Fix bug where `Search` filter was not displaying with the first query.
+
+## 0.16.1
+
+* Fix build that was improperly pushed to npm in 0.16.0.
+
+## 0.16.0
+
+* Update search-ui, add loader, increase debounce length, and reset search when modal is closed for `Search` component. [#145](https://github.com/mapbox/dr-ui/pull/145)
+
+## 0.15.2
+
+* Fix build to remove `optionalDependencies` and `jest` from package.json. [#143](https://github.com/mapbox/dr-ui/pull/143)
+
+## 0.15.1
+
+* Fix toggle in `Search` to be a button and track toggle event.
+
+## 0.15.0
+
+* Fix z-index issue with `Search` modal and `TopbarSticker`. [#139](https://github.com/mapbox/dr-ui/pull/139)
 * Add ability to filter `Search` results by current site or all docs. [#138](https://github.com/mapbox/dr-ui/pull/138)
+* Add Segment events for tracking queries and clicks in the `Search` component. [#140](https://github.com/mapbox/dr-ui/pull/14)
 
 ## 0.14.0